Create CVE-2019-12986.yaml
The trace_route action in DiagnosticsController is susceptible to command injection by a remote, unauthenticated attacker. Specifically, the trace_route function does not sufficiently validate or sanitize HTTP request parameter values that are used to construct a shell command. An attacker can trigger this vulnerability by routing traffic through the Collector controller and supplying a crafted value for ipAddress.

Signed-off-by: GwanYeong Kim <gy741.kim@gmail.com>