Update zip-backup-files.yaml new file name added
Template Validation
I've validated this template locally?
 YES
Additional Details (leave it blank if not applicable)

This template is trying to find backup files. In the "filename" part, I added the "inetpub". So, now it can find inetpub.zip, inetpub.rar etc..
I recently find an inetpub.zip in a website so I think, It ll be usefull.